+++ Sources/Heeler/Demo/DemoScreenshotMode.swift
@@ -444,6 +444,7 @@
 
         func closePane(_ params: PaneTarget) async throws {}
         func renameAgent(_ params: AgentRenameParams) async throws {}
+        func renamePane(_ params: PaneRenameParams) async throws {}
         func renameWorkspace(_ params: WorkspaceRenameParams) async throws {}
 
         func subscribeToEvents(